requirements.txt
pandas
numpy
scikit-learn==1.6.1
matplotlib
dotenv
seaborn
pymysql
cryptography
hydra-core
pandera[pandas]